src/main/java/no/kristiania/library/LibraryServer.java
package no.kristiania.library;

import org.eclipse.jetty.server.Server;
import org.eclipse.jetty.servlet.ServletContextHandler;

import javax.sql.DataSource;
import java.net.MalformedURLException;
import java.net.URL;

public class LibraryServer {
    private final Server server;

    public LibraryServer(int port, DataSource dataSource) {
        this.server = new Server(port);
        var handler = new ServletContextHandler();
        handler.setContextPath("/");
        server.setHandler(handler);
    }

    public URL getURL() throws MalformedURLException {
        return server.getURI().toURL();
    }

    public void start() throws Exception {
        server.start();
    }
}
